Remove ChefDK installation; Fetch SSL certificates
Having the delivery-truck and delivery_builder cookbooks both try to
install ChefDK (at different versions) resulted in them continuously
overwriting each other. delivery_builder has been updated to the latest
ChefDK and effort will be made to keep that up to date.

To allow the publish recipe to talk to a Chef Server with an unsigned
SSL cert, we'll need to fetch that cert using `knife ssl fetch`.