Publish Phase v1
This is the first iteration as a generic publish phase that will allow
the user to a) push their project repo to a github and b) upload any
cookbooks in that repo to the Delivery Chef Server.

This change also contains a rudamentary implementation of secret
management. The details of this are outlined in the README.

Also included in this change is a small helper script (sandbox) that I use
to quickly switch which Delivery server I am pointing to. By running
this script I can switch between submit reviews to my local dev or the
live environment. This helps speed up my local process. It is not
intended for external consumption and as such is not documented.