Deploy phase triggers CCRs on appropriate targets
As a user, I would like the delivery truck deploy phase recipe to trigger CCRs on nodes within target environments.

We have copied over the delivery_builder_deploy resource/provider from delivery_builder cookbook and overwritten them as a single-file LWRP library. We have writte them as a libary and not as LWRPs so that they could be leveraged inside a non-cookbook (i.e. if we decide to distribute those resources in a gem like we do with chef-provisioning). We've writte them in a single file because that is the pattern we are moving towards with the `crazytown` work. It also reduces the number of files.