Ensure that version was updated in verify/syntax
delivery-truck will upload your cookbook to the Chef Server for you but
it does so while also freezing it (via --freeze). If you submit a change
to your cookbook but don't update the metadata then you will fail when
it gets to this step (after you've already merged to master).

This change will simply raise an error if you forget to bump the version
in your metadata file.

Closes opscode-cookbooks/delivery-truck#1