Only enforce version bump on cookbook-related files
There may be times when you want to change non-cookbook related files like the
README, tests or some other root files. Updates to these files wouldn't (or shouldn't)
require a bump in the version number. The check in verify:syntax now only throws
an error if you modify a file in `recipes`, `libraries`, `attributes`, `files` or
`templates` and neglect to change the version number.