Add `compile_time` to `chef_gem` resources
As per https://github.com/chef/chef/blob/master/lib/chef/resource/chef_gem.rb#L49-L64
we have to specify `compile_time false` if we don't want the gems to be installed on
the compile phase. If we don't set this we will try to install them on both, the compile
and execution phases.