Add workaround for RuboCop 0.35.0 `Etc.getlogin` bug
If a builder executes the `rubocop` command with the latest version the
job fails with a `no implicit conversion of nil into String` error. This
is a known issue with this version and has a fix pending:
https://github.com/bbatsov/rubocop/issues/2407
https://github.com/bbatsov/rubocop/pull/2408

The easiest delivery-truck workaround is to ensure $USER is set in the
subprocess that executes the linting.