Added ability to override default of delivery_knife_rb as supermarket creds.
When uploading to the supermarket in publish, being required to use delivery_knife_rb is very limiting.
A user can now override those default creds with the use of the publish option supermarket-custom-credentials = true.
The user must upload supermarket_user and supermarket_key to the secrets data bag, otherwise a relevant error will be thrown.