bumped_version? checks if cookbook version changed
Previously, bumped_version? checks that if a cookbook in a change request
has been modified, then its metadata.(rb|json) has also been modified.
Unfortunately, this does not actually verify if there has been a
cookbook version bump and if the bump was sufficient after merging.

This change fixes this by actually parsing out the versions in the
metadata.rb files. It requires the corresponding change in
delivery-sugar that provides the API in DeliverySugar::Change to parse
cookbook versions.