Reduce the footprint of acceptance -> union race condition
There is a race condition in all stage transitions because
we pull version pinnings from the previous env, which may
be untested/unapproved.  It's especially an issue when a stage
is re-run because that provides a large time window for changes
to accumulate.  To mitigate that this change only pulls versions from
acceptance to union when union has not previously seen that change id.

It is not perfect and has some race conditions itself, but it decreases
the footprint.  A full fix will be much more invasive.