Do not ignore rules when cookbook will be share to Supermarket
By default we ignore a couple of rules that we we don't need to ignore
if the cookbook will be share to Supermarket.

Even though you are not sharing your cookbook to Supermarket and you
don't want to ignore those rules you can set `ignore_rules` to be empty
and we will not ignore them.