Change Release Process - Stage 1
We are about to change the way we release delivery-truck to our
customers, this involves two different Stage.

In this Stage we will share delivery-truck to Supermarket (Actual
Release) but also we will push to master in Github, this way we have a
period of time where we have backwards compatibility for anyone that is
still pulling from Github. (This will change in Stage 2)

During this time frame if you are a Delivery User consuming delivery-truck
you have to remove the following entries from your `Berksfile`:
```
cookbook 'delivery-truck',
  git: 'https://github.com/chef-cookbooks/delivery-truck.git',
  branch: 'master'

cookbook 'delivery-sugar',
  git: 'https://github.com/chef-cookbooks/delivery-sugar.git',
  branch: 'master'
```

This will make your build-cookbook to consume `delivery-truck` from
Supermarket instead of Github

We also need to eat our own dogfood of delivery-sugar coming from DCC,
so when this cookbook gets new changes we immediately pick them here.