Verify cookbook_version exist before promoting
When there are `project_artifacts` that are not pinned in the
`cookbook_version` object we have some project that are trying to get
pinned to `nil`. That will return a 400 Bad Response from the Chef
Server because you can't do that.

This change verifies that the project_artifacts has a cookbook_version,
if it doesn't it will notify the end-user with a WARNING.