Add support for cookstyle linter
This change supports an opt-in feature in the `lint`
stage to use `cookstyle` instead of `rubocop` for
linting.

The user will set this using the `enable_cookstyle`
configuration setting in `config.json`, which will cause
cookstyle tests to be executed against any changed
cookbooks.

If `cookstyle` is not available on the builder/runner,
these tests will not occur.

This change also updated the README with instructions
on how to use this new functionality.

This change also adds the `live_stream` parameter
to the `rubocop` tests for more output to the
console.

Signed-off-by: Matt Stratton <matt.stratton@gmail.com>